#136 - placing classes in more sensible packages
