+++ exploration-module/src/test/java/io/edpn/backend/exploration/application/service/SystemEliteIdUpdateEventListenerTest.java
@@ -1,20 +1,18 @@
-package io.edpn.backend.exploration.adapter.applicationevent;
+package io.edpn.backend.exploration.application.service;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.edpn.backend.exploration.application.domain.Message;
 import io.edpn.backend.exploration.application.domain.System;
-import io.edpn.backend.exploration.application.domain.SystemCoordinateRequest;
-import io.edpn.backend.exploration.application.domain.SystemCoordinateUpdatedEvent;
-import io.edpn.backend.exploration.application.dto.persistence.entity.mapper.SystemCoordinatesResponseMapper;
+import io.edpn.backend.exploration.application.domain.SystemEliteIdRequest;
+import io.edpn.backend.exploration.application.domain.SystemEliteIdUpdatedEvent;
+import io.edpn.backend.exploration.application.dto.persistence.entity.mapper.SystemEliteIdResponseMapper;
 import io.edpn.backend.exploration.application.dto.web.object.MessageDto;
 import io.edpn.backend.exploration.application.dto.web.object.mapper.MessageDtoMapper;
 import io.edpn.backend.exploration.application.port.outgoing.message.SendMessagePort;
 import io.edpn.backend.exploration.application.port.outgoing.system.LoadSystemPort;
-import io.edpn.backend.exploration.application.port.outgoing.systemcoordinaterequest.DeleteSystemCoordinateRequestPort;
-import io.edpn.backend.exploration.application.port.outgoing.systemcoordinaterequest.LoadSystemCoordinateRequestBySystemNamePort;
-import io.edpn.backend.exploration.application.port.outgoing.systemcoordinaterequest.SystemCoordinatesUpdatedEventListener;
-import io.edpn.backend.exploration.application.service.SystemEliteIdInterModuleCommunicationService;
-import io.edpn.backend.messageprocessorlib.application.dto.eddn.data.SystemCoordinatesResponse;
+import io.edpn.backend.exploration.application.port.outgoing.systemeliteidrequest.DeleteSystemEliteIdRequestPort;
+import io.edpn.backend.exploration.application.port.outgoing.systemeliteidrequest.LoadSystemEliteIdRequestBySystemNamePort;
+import io.edpn.backend.messageprocessorlib.application.dto.eddn.data.SystemEliteIdResponse;
 import io.edpn.backend.util.Module;
 import lombok.SneakyThrows;
 import org.junit.jupiter.api.BeforeEach;
@@ -41,18 +39,18 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
-class SystemCoordinateUpdatedEventListenerTest {
+class SystemEliteIdUpdateEventListenerTest {
 
     @Mock
     private LoadSystemPort loadSystemPort;
     @Mock
-    private LoadSystemCoordinateRequestBySystemNamePort loadSystemCoordinateRequestBySystemNamePort;
+    private LoadSystemEliteIdRequestBySystemNamePort loadSystemEliteIdRequestBySystemNamePort;
     @Mock
-    private DeleteSystemCoordinateRequestPort deleteSystemCoordinateRequestPort;
+    private DeleteSystemEliteIdRequestPort deleteSystemEliteIdRequestPort;
     @Mock
     private SendMessagePort sendMessagePort;
     @Mock
-    private SystemCoordinatesResponseMapper systemCoordinatesResponseMapper;
+    private SystemEliteIdResponseMapper systemEliteIdResponseMapper;
     @Mock
     private MessageDtoMapper messageMapper;
     @Mock
@@ -62,16 +60,16 @@ class SystemCoordinateUpdatedEventListenerTest {
     @Mock
     private ExecutorService executorService;
 
-    private SystemCoordinatesUpdatedEventListener underTest;
+    private SystemEliteIdEventListenerService underTest;
 
     @BeforeEach
     void setUp() {
-        underTest = new SystemCoordinatesEventListener(
+        underTest = new SystemEliteIdEventListenerService(
                 loadSystemPort,
-                loadSystemCoordinateRequestBySystemNamePort,
-                deleteSystemCoordinateRequestPort,
+                loadSystemEliteIdRequestBySystemNamePort,
+                deleteSystemEliteIdRequestPort,
                 sendMessagePort,
-                systemCoordinatesResponseMapper,
+                systemEliteIdResponseMapper,
                 messageMapper,
                 objectMapper,
                 retryTemplate,
@@ -82,41 +80,41 @@ class SystemCoordinateUpdatedEventListenerTest {
     @Test
     void onEvent_shouldProcessPendingRequest() {
         String systemName = "systemName";
-        SystemCoordinateUpdatedEvent systemCoordinateUpdatedEvent = new SystemCoordinateUpdatedEvent(mock(SystemEliteIdInterModuleCommunicationService.class), systemName);
+        SystemEliteIdUpdatedEvent systemEliteIdUpdatedEvent = new SystemEliteIdUpdatedEvent(mock(SystemEliteIdInterModuleCommunicationService.class), systemName);
         Module module = mock(Module.class);
         when(module.getName()).thenReturn("module");
-        SystemCoordinateRequest request1 = mock(SystemCoordinateRequest.class);
+        SystemEliteIdRequest request1 = mock(SystemEliteIdRequest.class);
         when(request1.requestingModule()).thenReturn(module);
-        when(loadSystemCoordinateRequestBySystemNamePort.loadByName(systemCoordinateUpdatedEvent.getSystemName())).thenReturn(List.of(request1));
+        when(loadSystemEliteIdRequestBySystemNamePort.loadByName(systemEliteIdUpdatedEvent.getSystemName())).thenReturn(List.of(request1));
         ArgumentCaptor<Runnable> runnableArgumentCaptor = ArgumentCaptor.forClass(Runnable.class);
         System mockSystem = mock(System.class);
         when(loadSystemPort.load(systemName)).thenReturn(Optional.of(mockSystem));
-        SystemCoordinatesResponse mockSystemCoordinatesResponse = mock(SystemCoordinatesResponse.class);
-        when(systemCoordinatesResponseMapper.map(mockSystem)).thenReturn(mockSystemCoordinatesResponse);
-        when(objectMapper.writeValueAsString(mockSystemCoordinatesResponse)).thenReturn("JSON_STRING");
-        Message coordinateKafkaMessage = new Message("module_systemCoordinatesResponse", "JSON_STRING");
+        SystemEliteIdResponse mockSystemEliteIdResponse = mock(SystemEliteIdResponse.class);
+        when(systemEliteIdResponseMapper.map(mockSystem)).thenReturn(mockSystemEliteIdResponse);
+        when(objectMapper.writeValueAsString(mockSystemEliteIdResponse)).thenReturn("JSON_STRING");
+        Message coordinateKafkaMessage = new Message("module_systemEliteIdResponse", "JSON_STRING");
         MessageDto coordinateMessageDto = mock(MessageDto.class);
         when(messageMapper.map(coordinateKafkaMessage)).thenReturn(coordinateMessageDto);
         when(sendMessagePort.send(coordinateMessageDto)).thenReturn(true);
         doAnswer(invocation -> ((RetryCallback<?, ?>) invocation.getArgument(0)).doWithRetry(null)).when(retryTemplate).execute(any());
 
-        underTest.onUpdatedEvent(systemCoordinateUpdatedEvent);
+        underTest.onUpdatedEvent(systemEliteIdUpdatedEvent);
 
         verify(executorService).submit(runnableArgumentCaptor.capture());
 
         // Verify runnable
         runnableArgumentCaptor.getAllValues().forEach(Runnable::run);
         verify(sendMessagePort).send(coordinateMessageDto);
-        verify(deleteSystemCoordinateRequestPort).delete(systemName, module);
+        verify(deleteSystemEliteIdRequestPort).delete(systemName, module);
     }
 
     @SneakyThrows
     @Test
     void onEvent_shouldThrowErrorWhenSystemNotFound() {
         String systemName = "systemName";
-        SystemCoordinateRequest request = mock(SystemCoordinateRequest.class);
-        SystemCoordinateUpdatedEvent event = new SystemCoordinateUpdatedEvent(mock(SystemEliteIdInterModuleCommunicationService.class), systemName);
-        when(loadSystemCoordinateRequestBySystemNamePort.loadByName(systemName)).thenReturn(List.of(request));
+        SystemEliteIdRequest request = mock(SystemEliteIdRequest.class);
+        SystemEliteIdUpdatedEvent event = new SystemEliteIdUpdatedEvent(mock(SystemEliteIdInterModuleCommunicationService.class), systemName);
+        when(loadSystemEliteIdRequestBySystemNamePort.loadByName(systemName)).thenReturn(List.of(request));
         when(loadSystemPort.load(systemName)).thenReturn(Optional.empty());
         ArgumentCaptor<Runnable> runnableArgumentCaptor = ArgumentCaptor.forClass(Runnable.class);
 
@@ -133,31 +131,31 @@ class SystemCoordinateUpdatedEventListenerTest {
     @Test
     void onEvent_shouldNotDeleteRequestWhenSendFails() {
         String systemName = "systemName";
-        SystemCoordinateUpdatedEvent systemCoordinateUpdatedEvent = new SystemCoordinateUpdatedEvent(mock(SystemEliteIdInterModuleCommunicationService.class), systemName);
-        SystemCoordinateRequest request1 = mock(SystemCoordinateRequest.class);
+        SystemEliteIdUpdatedEvent systemEliteIdUpdatedEvent = new SystemEliteIdUpdatedEvent(mock(SystemEliteIdInterModuleCommunicationService.class), systemName);
+        SystemEliteIdRequest request1 = mock(SystemEliteIdRequest.class);
         Module module = mock(Module.class);
         when(module.getName()).thenReturn("module");
         when(request1.requestingModule()).thenReturn(module);
-        when(loadSystemCoordinateRequestBySystemNamePort.loadByName(systemCoordinateUpdatedEvent.getSystemName())).thenReturn(List.of(request1));
+        when(loadSystemEliteIdRequestBySystemNamePort.loadByName(systemEliteIdUpdatedEvent.getSystemName())).thenReturn(List.of(request1));
         ArgumentCaptor<Runnable> runnableArgumentCaptor = ArgumentCaptor.forClass(Runnable.class);
         System mockSystem = mock(System.class);
         when(loadSystemPort.load(systemName)).thenReturn(Optional.of(mockSystem));
-        SystemCoordinatesResponse mockSystemCoordinatesResponse = mock(SystemCoordinatesResponse.class);
-        when(systemCoordinatesResponseMapper.map(mockSystem)).thenReturn(mockSystemCoordinatesResponse);
-        when(objectMapper.writeValueAsString(mockSystemCoordinatesResponse)).thenReturn("JSON_STRING");
-        Message coordinateKafkaMessage = new Message("module_systemCoordinatesResponse", "JSON_STRING");
+        SystemEliteIdResponse mockSystemEliteIdResponse = mock(SystemEliteIdResponse.class);
+        when(systemEliteIdResponseMapper.map(mockSystem)).thenReturn(mockSystemEliteIdResponse);
+        when(objectMapper.writeValueAsString(mockSystemEliteIdResponse)).thenReturn("JSON_STRING");
+        Message coordinateKafkaMessage = new Message("module_systemEliteIdResponse", "JSON_STRING");
         MessageDto coordinateMessageDto = mock(MessageDto.class);
         when(messageMapper.map(coordinateKafkaMessage)).thenReturn(coordinateMessageDto);
         when(sendMessagePort.send(coordinateMessageDto)).thenReturn(false);
         doAnswer(invocation -> ((RetryCallback<?, ?>) invocation.getArgument(0)).doWithRetry(null)).when(retryTemplate).execute(any());
 
-        underTest.onUpdatedEvent(systemCoordinateUpdatedEvent);
+        underTest.onUpdatedEvent(systemEliteIdUpdatedEvent);
 
         verify(executorService).submit(runnableArgumentCaptor.capture());
 
         // Verify runnable
         runnableArgumentCaptor.getAllValues().forEach(Runnable::run);
         verify(sendMessagePort).send(coordinateMessageDto);
-        verify(deleteSystemCoordinateRequestPort, never()).delete(systemName, module);
+        verify(deleteSystemEliteIdRequestPort, never()).delete(systemName, module);
     }
 }
